refactor: 사용자 프로필 이미지 사진 선택 코드
